Working on timing issue for ViewByManager
